feat: change the state of the active book
for this we create the active book reducer that take the action
triggered by event and generate by an action creator. This action
is dispatched to all reducers. This reducer takes a look at the
action.type and if 'BOOK_SELECTED' then return the active book.

Next we will combine this reducer to the rootReducer that generates
the global state.